requirements.txt
Click==7.0
dingding==0.5
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
PyYAML==5.1.2
Werkzeug==0.15.5
